✅ examples

+++ Exercises/501-1000/751-775/760ScoreAfterFlippingMat/scoreAfterFlippingMat.cpp
@@ -100,7 +100,7 @@ class Solution {
         cpy.push_back(grid[i]);
       }
     }
-    int halfsz = h/2;
+    float halfsz = (float)h/2;
     for (int i = 0; i < w; ++i) {
       if (count_c(cpy, i) < halfsz) {
         cpy = flipc(cpy, i);
